refactor: :truck: Moved the diles for the sidebar
